Comenzando ATDD Creado el test de aceptacion que no pasa al no estar la logica implementada
